requirements_test.txt
pylint==3.0.2
coverage==7.3.2
pytest-cov==4.1.0
pytest==7.4.3
pytest-html==4.0.2

python-dotenv==0.21.0
flask-restx==1.2.0
Flask-SQLAlchemy==3.1.1
pyodbc==5.0.1
requests==2.31.0